+++ lib/pages/blurting_tab/blurting.dart
@@ -15,6 +15,7 @@ import 'package:blurting/utils/util_widget.dart';
 import 'package:blurting/pages/blurting_tab/matching_ani.dart';
 import 'package:blurting/pages/blurting_tab/day_ani.dart';
 import 'package:http/http.dart' as http;
+import 'package:blurting/styles/styles.dart';
 
 /** */
 DateTime createdAt = DateTime.now();
