+++ lib/pages/signup_questions/mbti/mbti.dart
@@ -8,6 +8,7 @@ import 'package:blurting/pages/signup_questions/utils.dart';
 import 'package:blurting/pages/signup_questions/personality.dart';
 import 'package:http/http.dart' as http;
 import 'package:blurting/config/app_config.dart';
+import 'package:blurting/styles/styles.dart';
 
 class MBTIPage extends StatefulWidget {
   final String selectedGender;
