+++ lib/pages/startpage/start_page.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
 import 'dart:ui';
-import 'package:blurting/Utils/provider.dart';
+import 'package:blurting/styles/styles.dart';
 import 'package:blurting/pages/signup_questions/phone_number.dart';
 import 'package:blurting/token.dart';
 import 'package:flutter/material.dart';
