+++ lib/pages/useguide/useguide_done.dart
@@ -1,4 +1,4 @@
-import 'package:blurting/Utils/provider.dart';
+import 'package:blurting/styles/styles.dart';
 import 'package:blurting/mainApp.dart';
 import 'package:flutter/material.dart';
 import 'package:blurting/utils/util_widget.dart';
