+++ lib/pages/whisper_tab/whisper.dart
@@ -13,6 +13,7 @@ import 'package:blurting/config/app_config.dart';
 import 'package:http/http.dart' as http;
 import 'package:blurting/pages/whisper_tab/profileCard.dart';
 import 'package:extended_image/extended_image.dart' hide MultipartFile;
+import 'package:blurting/styles/styles.dart';
 
 class Whisper extends StatefulWidget {
   final String userName;
@@ -154,15 +155,14 @@ class _Whisper extends State<Whisper> {
                 chatMessages.insert(
                     i,
                     MyChat(
-                      index: 0,
-                      answerID: 0,
-                      message: widget.message,
-                      createdAt: widget.createdAt,
-                      read: true,
-                      isBlurting: false,
-                      likedNum: 0,
-                      event: false
-                    ));
+                        index: 0,
+                        answerID: 0,
+                        message: widget.message,
+                        createdAt: widget.createdAt,
+                        read: true,
+                        isBlurting: false,
+                        likedNum: 0,
+                        event: false));
               });
             }
           }
@@ -645,15 +645,14 @@ class _Whisper extends State<Whisper> {
 
     // 입력한 내용을 ListTile에 추가
     Widget newAnswer = MyChat(
-      index: 0,
-      answerID: 0,
-      message: message,
-      createdAt: '전송 중...',
-      read: true,
-      isBlurting: false,
-      likedNum: 0,
-      event: false
-    );
+        index: 0,
+        answerID: 0,
+        message: message,
+        createdAt: '전송 중...',
+        read: true,
+        isBlurting: false,
+        likedNum: 0,
+        event: false);
 
     // 소켓 서버에 데이터 전송
     if (message.isNotEmpty) {
@@ -739,16 +738,15 @@ class _Whisper extends State<Whisper> {
               if (chatData['userId'] ==
                   Provider.of<UserProvider>(context, listen: false).userId) {
                 fetchChatList = MyChat(
-                  index: 0,
-                  answerID: 0,
-                  message: chatData['chat'] as String? ?? '',
-                  createdAt: dateFormatAA.format(
-                      _parseDateTime(chatData['createdAt'] as String? ?? '')),
-                  read: read, // http에서 받아오는 거니까..
-                  isBlurting: false,
-                  likedNum: 0,
-                  event: false
-                );
+                    index: 0,
+                    answerID: 0,
+                    message: chatData['chat'] as String? ?? '',
+                    createdAt: dateFormatAA.format(
+                        _parseDateTime(chatData['createdAt'] as String? ?? '')),
+                    read: read, // http에서 받아오는 거니까..
+                    isBlurting: false,
+                    likedNum: 0,
+                    event: false);
               } else {
                 fetchChatList = OtherChat(
                     message: chatData['chat'] as String? ?? '',
